+++ src/common/constants.js
@@ -15,8 +15,8 @@ export const REDIRECT_DOMAINS = Object.freeze([
 ]);
 
 export const QS = Object.freeze({
-  VERSION: '1.0.4',
-  REVISION: 'R5',
+  VERSION: '1.0.5',
+  REVISION: 'R6',
   PLATFORM: 'Chrome',
   PACKAGE_NAME: 'quietshield.chrome',
   RULESETS: Object.freeze(['qs_ads','qs_trackers','qs_security','qs_redirects','qs_upgrade']),